% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bgschol_list}
\alias{bgschol_list}
\title{Get a list of tables}
\usage{
bgschol_list(con = NULL)
}
\arguments{
\item{con}{bgschol_con() object}
}
\description{
List of tables for big scholarly dataset.
}
\examples{
\dontrun{
# Connect
my_con <- my_con <- bqschol::bgschol_con(dataset = "cr_history")
# Get a list of tables
bgschol_list(my_con)
}

}
\seealso{
Other tables: 
\code{\link{bgschol_execute}()},
\code{\link{bgschol_query}()},
\code{\link{bgschol_tbl}()}
}
\concept{tables}
