% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bgschol_tbl}
\alias{bgschol_tbl}
\title{Access a single table with dpylr}
\usage{
bgschol_tbl(con = NULL, table = NULL)
}
\arguments{
\item{con}{bgschol_con connection (\code{\link[=bgschol_con]{bgschol_con()}})}

\item{table}{Database table}
}
\description{
dplyr interface treats tables as if they are in-memory data frames.
This enables accessing  big scholarly dataset without SQL.
}
\examples{
\dontrun{
my_con <- bqschol::bgschol_con()
cr_instant_df <- bgschol_tbl(my_con, table = "snapshot")
cr_instant_df \%>\%
    #top publishers
    dplyr::group_by(publisher) \%>\%
    dplyr::summarise(n = dplyr::n_distinct(doi)) \%>\%
    dplyr::arrange(desc(n)) \%>\%
    # Compute query and download to local session
    dplyr::collect()
   }
}
