% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqschol-package.R
\docType{package}
\name{bqschol-package}
\alias{bqschol}
\alias{bqschol-package}
\title{bqschol: Access and analyse SUB Göttingen Big Scholarly Data}
\description{
Database-specific package for various big scholarly 
    data on Google BigQuery curated at the SUB Göttingen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njahn82/bqschol}
  \item Report bugs at \url{https://github.com/njahn82/bqschol/issues}
}

}
\author{
\strong{Maintainer}: Najko Jahn \email{najko.jahn@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0001-5105-1463}{ORCID})

}
\keyword{internal}
