% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{bgschol_con}
\alias{bgschol_con}
\title{Create connection}
\usage{
bgschol_con(dataset = "cr_instant", path = NULL)
}
\arguments{
\item{dataset}{big scholarly datasets. Currently the
following datasets are availabe:
\itemize{
\item \emph{cr_instant} Most recent monthly Crossref metadata snapshot,
comprising metadata about journal articles published since 2008.
\item \emph{cr_history} Historic Crossref metadata snapshots starting Apr
2018.
\item \emph{oadoi_full} Unpaywall metadata since 2008.
}}

\item{path}{Path to JSON identifying the associated service account.}
}
\description{
Authorization is only possible through a service account token.
The toke can be requested from Najko
\url{mailto:najko.jahn@sub.uni-goettingen.de}
}
