% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bgschol_execute}
\alias{bgschol_execute}
\title{Execute a SQL query on the database}
\usage{
bgschol_execute(con = NULL, query = NULL)
}
\arguments{
\item{con}{bgschol_con connection (\code{\link[=bgschol_con]{bgschol_con()}})}

\item{query}{SQL string}
}
\description{
Returns a scalar numeric that specifies the number of rows
affected by the statement.
}
\details{
Wrapper for DBI::dbGetQuery
}
