% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bgschol_list}
\alias{bgschol_list}
\title{Get a list of tables}
\usage{
bgschol_list(con = NULL)
}
\arguments{
\item{con}{bgschol_con connection (\code{\link[=bgschol_con]{bgschol_con()}})}
}
\description{
List of tables for big scholarly dataset.
}
\examples{
\dontrun{
# Connect
my_con <- my_con <- bqschol::bgschol_con(dataset = "cr_history")
# Get a list of tables
bgschol_list(my_con)
}
}
