% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{bgschol_query}
\alias{bgschol_query}
\title{Perform a SQL query and retrieve result}
\usage{
bgschol_query(con = NULL, query = NULL)
}
\arguments{
\item{con}{bgschol_con connection (\code{\link[=bgschol_con]{bgschol_con()}})}

\item{query}{SQL string}
}
\description{
Wrapper for DBI::dbGetQuery
}
