% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqschol-package.R
\docType{package}
\name{bqschol-package}
\alias{bqschol}
\alias{bqschol-package}
\title{bqschol: Access SUB Göttingen Big Scholarly Data}
\description{
Database client for SUB Göttingen Big Scholarly Data
	stored on Google Big Query.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njahn82/bqschol}
  \item Report bugs at \url{https://github.com/njahn82/bqschol/issues}
}

}
\author{
\strong{Maintainer}: Najko Jahn \email{najko.jahn@sub.uni-goettingen.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
